chore: move reconnect to connection
