+++ blocks/lists.ts
@@ -232,7 +232,7 @@ const LISTS_CREATE_WITH = {
     this.updateShape_();
     // Reconnect any child blocks.
     for (let i = 0; i < this.itemCount_; i++) {
-      MutatorIcon.reconnect(connections[i], this, 'ADD' + i);
+      connections[i]?.reconnect(this, 'ADD' + i);
     }
   },
   /**
